Extracted PlanSupportCheckerTests to a separate test
