+++ libs/iap/src/test/kotlin/com/woocommerce/android/iap/pub/IAPPurchaseWPComPlanActionsTest.kt
@@ -1,9 +1,7 @@
 package com.woocommerce.android.iap.pub
 
 import com.android.billingclient.api.BillingClient
-import com.android.billingclient.api.BillingClientStateListener
 import com.android.billingclient.api.BillingResult
-import com.android.billingclient.api.ProductDetailsResult
 import com.android.billingclient.api.Purchase
 import com.android.billingclient.api.PurchasesResult
 import com.woocommerce.android.iap.internal.core.IAPBillingClientProvider
@@ -15,14 +13,11 @@ import com.woocommerce.android.iap.internal.core.IAPManager
 import com.woocommerce.android.iap.internal.core.IAPOutMapper
 import com.woocommerce.android.iap.internal.core.IAPPeriodicPurchaseStatusChecker
 import com.woocommerce.android.iap.internal.core.IAPPurchasesUpdatedListener
-import com.woocommerce.android.iap.internal.core.isSuccess
-import com.woocommerce.android.iap.internal.model.IAPProduct
 import com.woocommerce.android.iap.internal.network.IAPMobilePayAPI
 import com.woocommerce.android.iap.internal.network.model.CreateAndConfirmOrderResponse
 import com.woocommerce.android.iap.internal.planpurchase.IAPPurchaseWPComPlanActionsImpl
 import com.woocommerce.android.iap.internal.planpurchase.IAPPurchaseWpComPlanHandler
 import com.woocommerce.android.iap.pub.model.IAPError
-import com.woocommerce.android.iap.pub.model.IAPSupportedResult
 import com.woocommerce.android.iap.pub.model.WPComIsPurchasedResult
 import com.woocommerce.android.iap.pub.model.WPComProductResult
 import com.woocommerce.android.iap.pub.model.WPComPurchaseResult
@@ -57,7 +52,7 @@ class IAPPurchaseWPComPlanActionsTest {
         on { activity }.thenReturn(mock())
     }
 
-    private val iapProduct: IAPProduct = IAPProduct.WPPremiumPlanTesting
+    private val testPreparationHelper = IAPTestPreparationHelper(billingClientMock)
 
     private lateinit var purchasesUpdatedListener: IAPPurchasesUpdatedListener
     private lateinit var sut: IAPPurchaseWPComPlanActionsImpl
@@ -68,70 +63,6 @@ class IAPPurchaseWPComPlanActionsTest {
     }
 
     //region Tests
-    @Test
-    fun `given product in NON USD currency, when iap support checked, then success with false returned`() = runTest {
-        // GIVEN
-        setupQueryProductDetails(
-            responseCode = BillingClient.BillingResponseCode.OK,
-            currency = "AED"
-        )
-
-        // WHEN
-        val result = sut.isIAPSupported()
-
-        // THEN
-        assertThat((result as IAPSupportedResult.Success).isSupported).isFalse()
-    }
-
-    @Test
-    fun `given product USD currency, when iap support checked, then success with true returned`() = runTest {
-        // GIVEN
-        setupQueryProductDetails(responseCode = BillingClient.BillingResponseCode.OK)
-
-        // WHEN
-        val result = sut.isIAPSupported()
-
-        // THEN
-        assertThat((result as IAPSupportedResult.Success).isSupported).isTrue()
-    }
-
-    @Test
-    fun `given product fetching error, when iap support checked, then error returned`() = runTest {
-        // GIVEN
-        val responseCode = BillingClient.BillingResponseCode.DEVELOPER_ERROR
-        val debugMessage = "debug message"
-
-        setupQueryProductDetails(
-            responseCode = responseCode,
-            debugMessage = debugMessage
-        )
-
-        // WHEN
-        val result = sut.isIAPSupported()
-
-        // THEN
-        assertThat((result as IAPSupportedResult.Error).errorType).isInstanceOf(
-            IAPError.Billing.DeveloperError::class.java
-        )
-        assertThat((result.errorType as IAPError.Billing).debugMessage).isEqualTo(debugMessage)
-    }
-
-    @Test
-    fun `given connection to service failed, when iap support checked, then error returned`() = runTest {
-        // GIVEN
-        val debugMessage = "debug message"
-        setupSut(buildBillingResult(BillingClient.BillingResponseCode.SERVICE_UNAVAILABLE, debugMessage))
-
-        // WHEN
-        val result = sut.isIAPSupported()
-
-        // THEN
-        assertThat((result as IAPSupportedResult.Error).errorType).isInstanceOf(
-            IAPError.Billing.ServiceUnavailable::class.java
-        )
-        assertThat((result.errorType as IAPError.Billing).debugMessage).isEqualTo(debugMessage)
-    }
-
     @Test
     fun `given success and product purchased, when iap wp com plan purchase checked, then true returned`() = runTest {
         // GIVEN
@@ -147,7 +78,10 @@ class IAPPurchaseWPComPlanActionsTest {
                 )
             )
         )
-        setupQueryProductDetails(responseCode = BillingClient.BillingResponseCode.OK, debugMessage = debugMessage)
+        testPreparationHelper.setupQueryProductDetails(
+            responseCode = BillingClient.BillingResponseCode.OK,
+            debugMessage = debugMessage
+        )
 
         // WHEN
         val result = sut.isWPComPlanPurchased()
@@ -190,7 +124,7 @@ class IAPPurchaseWPComPlanActionsTest {
                 )
             )
 
-            setupQueryProductDetails(
+            testPreparationHelper.setupQueryProductDetails(
                 responseCode = responseCode,
                 productId = "another_product_id"
             )
@@ -219,7 +153,7 @@ class IAPPurchaseWPComPlanActionsTest {
                 )
             )
 
-            setupQueryProductDetails(responseCode = responseCode)
+            testPreparationHelper.setupQueryProductDetails(responseCode = responseCode)
 
             // WHEN
             val result = sut.isWPComPlanPurchased()
@@ -261,7 +195,7 @@ class IAPPurchaseWPComPlanActionsTest {
         val price = 100L
         val title = "wc_plan"
         val description = "best plan ever"
-        setupQueryProductDetails(
+        testPreparationHelper.setupQueryProductDetails(
             responseCode = responseCode,
             priceMicroCents = price,
             title = title,
@@ -283,7 +217,7 @@ class IAPPurchaseWPComPlanActionsTest {
     fun `given product query error, when fetching wp com plan product, then error returned`() = runTest {
         // GIVEN
         val debugMessage = "debug message"
-        setupQueryProductDetails(
+        testPreparationHelper.setupQueryProductDetails(
             responseCode = BillingClient.BillingResponseCode.FEATURE_NOT_SUPPORTED,
             debugMessage = debugMessage
         )
@@ -332,7 +266,7 @@ class IAPPurchaseWPComPlanActionsTest {
         val purchaseToken = "purchaseToken"
         val responseCode = BillingClient.BillingResponseCode.OK
 
-        setupQueryProductDetails(responseCode)
+        testPreparationHelper.setupQueryProductDetails(responseCode)
 
         setupMobilePayAPIMock(
             purchaseToken = purchaseToken,
@@ -389,7 +323,7 @@ class IAPPurchaseWPComPlanActionsTest {
         )
 
         val debugMessage = "debug message"
-        setupQueryProductDetails(
+        testPreparationHelper.setupQueryProductDetails(
             responseCode = BillingClient.BillingResponseCode.SERVICE_DISCONNECTED,
             debugMessage = debugMessage
         )
@@ -414,7 +348,7 @@ class IAPPurchaseWPComPlanActionsTest {
             listOf()
         )
 
-        setupQueryProductDetails(responseCode = BillingClient.BillingResponseCode.OK)
+        testPreparationHelper.setupQueryProductDetails(responseCode = BillingClient.BillingResponseCode.OK)
 
         val debugMessage = "debug message"
 
@@ -442,7 +376,7 @@ class IAPPurchaseWPComPlanActionsTest {
             listOf()
         )
 
-        setupQueryProductDetails(responseCode = BillingClient.BillingResponseCode.OK)
+        testPreparationHelper.setupQueryProductDetails(responseCode = BillingClient.BillingResponseCode.OK)
 
         // WHEN
         awaitAll(
@@ -473,7 +407,7 @@ class IAPPurchaseWPComPlanActionsTest {
 
             val responseCode = BillingClient.BillingResponseCode.OK
 
-            setupQueryProductDetails(responseCode)
+            testPreparationHelper.setupQueryProductDetails(responseCode)
 
             setupMobilePayAPIMock(
                 purchaseToken = purchaseToken,
@@ -517,7 +451,7 @@ class IAPPurchaseWPComPlanActionsTest {
         val purchaseToken = "purchaseToken"
         val responseCode = BillingClient.BillingResponseCode.OK
 
-        setupQueryProductDetails(responseCode)
+        testPreparationHelper.setupQueryProductDetails(responseCode)
 
         setupMobilePayAPIMock(
             purchaseToken = purchaseToken,
@@ -560,7 +494,7 @@ class IAPPurchaseWPComPlanActionsTest {
         val purchaseToken = "purchaseToken"
         val responseCode = BillingClient.BillingResponseCode.OK
 
-        setupQueryProductDetails(responseCode)
+        testPreparationHelper.setupQueryProductDetails(responseCode)
 
         setupMobilePayAPIMock(
             purchaseToken = purchaseToken,
@@ -603,7 +537,7 @@ class IAPPurchaseWPComPlanActionsTest {
 
             val responseCode = BillingClient.BillingResponseCode.OK
 
-            setupQueryProductDetails(responseCode)
+            testPreparationHelper.setupQueryProductDetails(responseCode)
 
             setupMobilePayAPIMock(
                 purchaseToken = purchaseToken,
@@ -650,7 +584,7 @@ class IAPPurchaseWPComPlanActionsTest {
 
             val responseCode = BillingClient.BillingResponseCode.OK
 
-            setupQueryProductDetails(responseCode)
+            testPreparationHelper.setupQueryProductDetails(responseCode)
 
             setupMobilePayAPIMock(
                 purchaseToken = purchaseToken,
@@ -714,7 +648,7 @@ class IAPPurchaseWPComPlanActionsTest {
 
             val responseCode = BillingClient.BillingResponseCode.OK
 
-            setupQueryProductDetails(responseCode)
+            testPreparationHelper.setupQueryProductDetails(responseCode)
 
             setupMobilePayAPIMock(
                 purchaseToken = purchaseToken,
@@ -747,7 +681,7 @@ class IAPPurchaseWPComPlanActionsTest {
 
             val responseCode = BillingClient.BillingResponseCode.OK
 
-            setupQueryProductDetails(responseCode)
+            testPreparationHelper.setupQueryProductDetails(responseCode)
 
             // WHEN
             sut.purchaseWPComPlan(activityWrapperMock)
@@ -774,7 +708,7 @@ class IAPPurchaseWPComPlanActionsTest {
 
             val responseCode = BillingClient.BillingResponseCode.OK
 
-            setupQueryProductDetails(responseCode)
+            testPreparationHelper.setupQueryProductDetails(responseCode)
 
             // WHEN
             sut.purchaseWPComPlan(activityWrapperMock)
@@ -782,7 +716,7 @@ class IAPPurchaseWPComPlanActionsTest {
             // THEN
             verifyNoInteractions(mobilePayAPIMock)
         }
-//endregion
+    //endregion
 
     private suspend fun setupPurchaseQuery(
         @BillingClient.BillingResponseCode responseCode: Int,
@@ -793,31 +727,6 @@ class IAPPurchaseWPComPlanActionsTest {
             PurchasesResult(
                 buildBillingResult(responseCode, debugMessage),
                 purchases
-            )
-        )
-    }
-
-    private suspend fun setupQueryProductDetails(
-        responseCode: Int,
-        debugMessage: String = "",
-        productId: String = iapProduct.productId,
-        priceMicroCents: Long = 10_000L,
-        currency: String = "USD",
-        title: String = "title",
-        description: String = "description"
-    ) {
-        val productDetails = buildProductDetails(
-            productId = productId,
-            name = "productName",
-            price = priceMicroCents,
-            currency = currency,
-            title = title,
-            description = description,
-        )
-        whenever(billingClientMock.queryProductDetails(any())).thenReturn(
-            ProductDetailsResult(
-                buildBillingResult(responseCode, debugMessage),
-                listOf(productDetails)
             )
         )
     }
@@ -848,32 +757,31 @@ class IAPPurchaseWPComPlanActionsTest {
             on { provideBillingClient() }.thenReturn(billingClientMock)
         }
         val outMapper = IAPOutMapper()
-        setupBillingClient(connectionResult)
+        testPreparationHelper.setupBillingClient(connectionResult)
         val iapBillingClientStateHandler = IAPBillingClientStateHandler(
             billingClientProvider,
             outMapper,
             logWrapperMock,
         )
-        val iapManager = buildIapManager(iapBillingClientStateHandler, purchasesUpdatedListener, outMapper)
+        val iapManager = IAPManager(
+            iapBillingClientStateHandler,
+            outMapper,
+            IAPInMapper(),
+            purchasesUpdatedListener,
+            billingFlowParamsBuilderMock,
+            periodicPurchaseStatusCheckerMock,
+            logWrapperMock,
+        )
         val purchaseWpComPlanHandler = IAPPurchaseWpComPlanHandler(
             iapMobilePayAPI = mobilePayAPIMock,
             iapManager = iapManager,
         )
         sut = IAPPurchaseWPComPlanActionsImpl(
             purchaseWpComPlanHandler = purchaseWpComPlanHandler,
-            iapManager = buildIapManager(iapBillingClientStateHandler, purchasesUpdatedListener, outMapper),
+            iapManager = iapManager,
             REMOTE_SITE_ID,
             iapProduct = iapProduct,
         )
-    }
-
-    private fun setupBillingClient(connectionResult: BillingResult) {
-        whenever(billingClientMock.startConnection(any())).thenAnswer {
-            val listener = it.arguments[0] as BillingClientStateListener
-            listener.onBillingSetupFinished(connectionResult)
-            whenever(billingClientMock.isReady).thenReturn(connectionResult.isSuccess)
-        }
-        whenever(billingClientMock.isReady).thenReturn(false)
     }
 
     private fun setupPeriodicJob(purchasesResult: PurchasesResult) {
@@ -884,18 +792,4 @@ class IAPPurchaseWPComPlanActionsTest {
                 mock<Job>()
             }
     }
-
-    private fun buildIapManager(
-        iapBillingClientStateHandler: IAPBillingClientStateHandler,
-        purchasesUpdatedListener: IAPPurchasesUpdatedListener,
-        outMapper: IAPOutMapper
-    ) = IAPManager(
-        iapBillingClientStateHandler,
-        outMapper,
-        IAPInMapper(),
-        purchasesUpdatedListener,
-        billingFlowParamsBuilderMock,
-        periodicPurchaseStatusCheckerMock,
-        logWrapperMock,
-    )
 }
